refactor

+++ alembic/versions/e5838d1ce582_initial.py
@@ -33,9 +33,9 @@ def upgrade() -> None:
     sa.Column('impact_event_id', sa.String(), nullable=False),
     sa.Column('asteroid_id', sa.String(), nullable=False),
     sa.Column('date', sa.String(), nullable=False),
-    sa.Column('impact_probability', sa.Float(), nullable=False),
+    sa.Column('impact_probability', sa.Numeric(), nullable=False),
     sa.Column('energy', sa.Numeric(precision=20, scale=2), nullable=False),
-    sa.Column('dangerous_score', sa.Float(), nullable=False),
+    sa.Column('dangerous_score', sa.Numeric(), nullable=False),
     sa.ForeignKeyConstraint(['asteroid_id'], ['asteroids.asteroid_id'], ),
     sa.PrimaryKeyConstraint('impact_event_id')
     )
